trying to debug and solve the errors and complet the code
