+++ models/ngspice/testing/regression/bjt_cj/device_netlists/npn_EBJ.spice
@@ -14,7 +14,7 @@
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
-Vb b 0 dc -3.3
+Vb b 0 dc 3.0
 
 .temp {{temp}}
 .options tnom={{temp}}
@@ -23,7 +23,7 @@ Vb b 0 dc -3.3
 xq1 0 b 0 0 {{device}}
 
 
-.DC Vb 0 -3.3 -0.1
+.DC Vb 0 3.0 0.1
 
 .control
