+++ models/ngspice/testing/regression/bjt_cj/models_regression.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Efabless Corporation
+# Copyright 2022 GlobalFoundries PDK Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,7 +20,6 @@ Usage:
   --num_cores=<num>      Number of cores to be used by simulator
 """
 
-from re import T
 from docopt import docopt
 import pandas as pd
 import numpy as np
@@ -28,160 +27,407 @@ import os
 from jinja2 import Template
 import concurrent.futures
 import shutil
-import warnings
+import multiprocessing as mp
 
-warnings.simplefilter(action="ignore", category=FutureWarning)
+import glob
+
+
+PASS_THRESH = 2.0  # threshold value for passing devices
+NO_ROWS_NPN = 54  # no.of combinations extracted from npn sheet
+NO_ROWS_PNP = 24  # no.of combinations extracted from pnp sheet
 
 
 def call_simulator(file_name):
     """Call simulation commands to perform simulation.
     Args:
-        file_name (str): Netlist file name.
+        file_name(str): Netlist file name.
     """
-    os.system(f"ngspice -b -a {file_name} -o {file_name}.log > {file_name}.log")
-
-
-def ext_measured(device, vn, d_in, cv_sim, corner, start, dirpath):
-
-    # Get dimensions used for each device
-    if "vnpn" in device:
-        loops = 3
-    else:
-        loops = 2
-
-    # Extracting measured values for each W & L
-    for i in range(start, loops + start):
-
-        # Special case for 1st measured values
-        if i == 0:
-            # measured Id
-            col_list = [f"{vn}", f"{d_in}_{corner}"]
-            df_measured = pd.read_csv(f"{dirpath}/{device}.csv", usecols=col_list)
-            df_measured.columns = [f"{vn}", f"{d_in}_{corner}"]
-            df_measured.to_csv(
-                f"{dirpath}/measured_{cv_sim}/{i-start}_measured_{device}.csv",
-                index=False,
-            )
-        else:
-            # measured Id
-            col_list = [f"{vn}", f"{d_in}_{corner}.{i}"]
-            df_measured = pd.read_csv(f"{dirpath}/{device}.csv", usecols=col_list)
-            df_measured.columns = [f"{vn}", f"{d_in}_{corner}"]
-            df_measured.to_csv(
-                f"{dirpath}/measured_{cv_sim}/{i-start}_measured_{device}.csv",
-                index=False,
-            )
-
-
-def ext_simulated(device, vn, d_in, cv_sim, corner, temp, dirpath, cap):
-
-    # Get dimensions used for each device
-    netlist_tmp = f"./device_netlists/{cap}.spice"
-
-    if "EBJ" in cap:
-        i = 0
-    elif "CBJ" in cap:
-        i = 1
-    else:
-        i = 2
-
-    with open(netlist_tmp) as f:
-        tmpl = Template(f.read())
-        os.makedirs(f"{dirpath}/{device}_netlists_{cv_sim}", exist_ok=True)
-        with open(
-            f"{dirpath}/{device}_netlists_{cv_sim}/{i}_{device}_netlist_{device}.spice",
-            "w",
-        ) as netlist:
-            netlist.write(
-                tmpl.render(device=device, i=i, temp=temp, cv_sim=cv_sim, corner=corner)
-            )
-        netlist_path = (
-            f"{dirpath}/{device}_netlists_{cv_sim}/{i}_{device}_netlist_{device}.spice"
-        )
-        # Running ngspice for each netlist
-        with concurrent.futures.ProcessPoolExecutor(
-            max_workers=workers_count
-        ) as executor:
-            executor.submit(call_simulator, netlist_path)
-
-        # Writing simulated data
-        df_simulated = pd.read_csv(
-            f"{dirpath}/simulated_{cv_sim}/{i}_simulated_{device}.csv",
-            header=None,
-            delimiter=r"\s+",
-        )
-        df_simulated.to_csv(
-            f"{dirpath}/simulated_{cv_sim}/{i}_simulated_{device}.csv", index=False
-        )
-        df_simulated.columns = [f"{vn}", f"{d_in}_{corner}"]
-        df_simulated.to_csv(
-            f"{dirpath}/simulated_{cv_sim}/{i}_simulated_{device}.csv", index=False
-        )
-
-
-def error_cal(device, vn, d_in, cv_sim, corner, temp, dirpath):
-
-    # Get dimensions used for each device
-    loops = 2
-    df_final = pd.DataFrame()
-    for i in range(0, loops):
-        if i == 0:
-            cap = "EBJ"
-        elif i == 1:
-            cap = "CBJ"
-        else:
-            cap = "CSJ"
-
-        measured = pd.read_csv(f"{dirpath}/measured_{cv_sim}/{i}_measured_{device}.csv")
-        simulated = pd.read_csv(
-            f"{dirpath}/simulated_{cv_sim}/{i}_simulated_{device}.csv"
-        )
-
-        error_1 = round(
-            100
-            * abs(
-                (abs(measured.iloc[:, 1]) - abs(simulated.iloc[:, 1]))
-                / abs(measured.iloc[:, 1])
-            ),
-            6,
-        )
-
-        df_error = pd.DataFrame(data=[measured.iloc[:, 0], error_1]).transpose()
-        df_error.to_csv(
-            f"{dirpath}/error_{cv_sim}/{i}_{device}_error_{device}.csv", index=False
-        )
-
-        # Mean error
-        mean_error = df_error[f"{d_in}_{corner}"].mean()
-        # Max error
-        max_error = df_error[f"{d_in}_{corner}"].max()
-        # Max error location
-        max_index = max((df_error == max_error).idxmax())
-        max_location_vgs = (df_error == max_error).idxmax(axis=1)[max_index]
-        max_location_vds = df_error[f"{vn}"][max_index]
-
-        df_final_ = {
-            "Run no.": f"{i}",
-            "Temp": f"{temp}",
-            "Device name": f"{device}",
-            "corner": f"{corner}",
-            "Junction": f"{cap}",
-            "Simulated_Val": f"{cv_sim}",
-            "Mean error%": f'{"{:.2f}".format(mean_error)}',
-            "Max error%": f'{"{:.2f}".format(max_error)} @ {max_location_vgs} & vn (V) = {max_location_vds}',
-        }
-        df_final = df_final.append(df_final_, ignore_index=True)
-    # Max mean error
-    print(df_final)
-    df_final.to_csv(f"{dirpath}/Final_report_{cv_sim}.csv", index=False)
-    out_report = pd.read_csv(f"{dirpath}/Final_report_{cv_sim}.csv")
-    print("\n", f"Max. mean error = {out_report['Mean error%'].max()}%")
-    print(
-        "====================================================================================================================================================="
+    return os.system(
+        f"ngspice -b -a {file_name} -o {file_name}.log > {file_name}.log"
     )
 
 
+def ext_measured(
+    cj_file: str, dev: str, devices: list, dev_path: str, no_rows=int
+) -> pd.DataFrame:
+    """Extracting the measured data of npn devices from excel sheet
+
+    Args:
+         cj_file(str): path to the data sheet
+         dev(str): device name whether npn or pnp
+         devices(list): list for undertest devices
+         dev_path(str): A path where extracted data is stored
+         no_rows(int): no.of combinations extracted from npn, pnp sheet
+
+    Returns:
+         df_measured(pd.DataFrame): A data frame contains all extracted data
+
+    """
+
+    # Reading excel sheet and creating data frame
+    df = pd.read_excel(cj_file)
+
+    # temp_range is threshold for switching between 25, -40, 125
+    temp_range = int(no_rows / 3)
+    # initiating empty list for appendindurationg dataframes
+    all_dfs = list()
+
+    # Extracting measured values for each Device
+    for i in range(no_rows):
+
+        # building up temperature
+        if i in range(0, temp_range):
+            temp = 25
+        elif i in range(temp_range, 2 * temp_range):
+            temp = -40
+        else:
+            temp = 175
+
+        # extracted columns from sheet
+        temp_value = list()
+        dev_names = list()
+        cj_measured = list()
+        cap_names = list()
+
+        # reading third column for getting device name, cap_name
+        data = df["Unnamed: 2"][i]  # data is a string now
+        cap_name = data[4:7]
+        end_location = data.find("(")
+        read_dev_name = data[8:end_location]
+        space = read_dev_name.find(" ")
+        read_dev_name = read_dev_name[0:space]
+
+        # renaming the device like the standard
+        # for npn devices:
+        if dev == "npn":
+            if read_dev_name == "10x10":
+                dev_name = devices[0]
+            elif read_dev_name == "5x5":
+                dev_name = devices[1]
+            elif read_dev_name == "0p54x16":
+                dev_name = devices[2]
+            elif read_dev_name == "0p54x8":
+                dev_name = devices[3]
+            elif read_dev_name == "0p54x4":
+                dev_name = devices[4]
+            elif read_dev_name == "0p54x2":
+                dev_name = devices[5]
+
+        # for pnp devices:
+        elif dev == "pnp":
+            if read_dev_name == "0p42x10":
+                dev_name = devices[0]
+            elif read_dev_name == "0p42x5":
+                dev_name = devices[1]
+            elif read_dev_name == "10x10":
+                dev_name = devices[2]
+            elif read_dev_name == "5x5":
+                dev_name = devices[3]
+
+        # extracting C-V measured data
+        # Special case for 1st measured values
+        if i == 0:
+            cj_values = df[
+                [
+                    "Vj",
+                    "bjt_typical",
+                    "bjt_ff",
+                    "bjt_ss"
+                ]
+            ].copy()
+
+            cj_values.rename(
+                columns={
+                    "Vj": "measured_volt",
+                    "bjt_typical": "measured_bjt_typical",
+                    "bjt_ff": "measured_bjt_ff",
+                    "bjt_ss": "measured_bjt_ss"
+                },
+                inplace=True
+            )
+
+        else:
+            cj_values = df[
+                [
+                    "Vj",
+                    f"bjt_typical.{i}",
+                    f"bjt_ff.{i}",
+                    f"bjt_ss.{i}"
+                ]
+            ].copy()
+
+            cj_values.rename(
+                columns={
+                    "Vj": "measured_volt",
+                    f"bjt_typical.{i}": "measured_bjt_typical",
+                    f"bjt_ff.{i}": "measured_bjt_ff",
+                    f"bjt_ss.{i}": "measured_bjt_ss"
+                },
+                inplace=True
+            )
+
+        os.makedirs(f"{dev_path}/cj_measured", exist_ok=True)
+        cj_values.dropna(axis=0, inplace=True)
+        cj_values.to_csv(
+            f"{dev_path}/cj_measured/measured_{dev_name}_t{temp}.csv",
+            index=False
+        )
+
+        dev_names.append(dev_name)
+        cap_names.append(cap_name)
+        temp_value.append(temp)
+        cj_measured.append(
+            f"{dev_path}/cj_measured/measured_{dev_name}_t{temp}.csv"
+        )
+
+        sdf = {
+            "device": dev_names,
+            "temp": temp_value,
+            "cap": cap_names,
+            "cj_measured": cj_measured
+        }
+        sdf = pd.DataFrame(sdf)
+        all_dfs.append(sdf)
+
+    df = pd.concat(all_dfs)
+    df.dropna(axis=0, inplace=True)
+    return df
+
+
+def run_sim(dirpath: str, cap: str, device: str, temp: float) -> dict:
+    """Run simulation at specific information and corner
+    Args:
+        dirpath(str): path to the file where we write data
+        cap(str): under-test cap to select proper netlist
+        device(str): the device instance will be simulated
+        temp(float): a specific temp for simulation
+
+    Returns:
+        info(dict): results are stored in,
+        and passed to the run_sims function to extract data
+    """
+
+    corners = ["typical", "ff", "ss"]
+    for corner in corners:
+        info = dict()
+        info["device"] = device
+        info["temp"] = temp
+        dev = device.split("_")[0]
+
+        netlist_tmp = f"device_netlists/{dev}_{cap}.spice"
+        temp_str = "{:.1f}".format(temp)
+        netlist_path =f"{dirpath}/{dev}_netlists/netlist_{device}_t{temp_str}_c{corner}.spice"
+
+        result_path = (
+            f"{dirpath}/cj_simulated/simulated_{device}_t{temp_str}_c{corner}.csv"
+        )
+
+        # initiating the directory in which results will be stored
+        os.makedirs(f"{dirpath}/cj_simulated", exist_ok=True)
+
+
+        with open(netlist_tmp) as f:
+            tmpl = Template(f.read())
+            os.makedirs(f"{dirpath}/{dev}_netlists", exist_ok=True)
+
+            with open(netlist_path, "w") as netlist:
+                netlist.write(
+                    tmpl.render(
+                        device=device,
+                        temp=temp_str,
+                        corner=corner
+                    )
+                )
+
+            with open(netlist_path, "r") as netlist:
+                netlist_str = netlist.read()
+            netlist_str.replace("endend", "end")
+            print(netlist_str)
+
+            with open(netlist_path, "w") as netlist:
+                    netlist.write(netlist_str)
+
+        # Running ngspice for each netlist
+        try:
+            call_simulator(netlist_path)
+
+            # check if results stored in csv file or not!
+            if os.path.exists(result_path):
+                bjt_simu_cj = result_path
+            else:
+                bjt_simu_cj = "None"
+
+        except Exception:
+            bjt_simu_cj = "None"
+
+    info["cj_simulated"]= bjt_simu_cj
+
+    return info
+
+
+def run_sims(
+    df: pd.DataFrame, dirpath: str, num_workers=mp.cpu_count()
+) :
+    """passing netlists to run_sim function
+        and storing the results csv files into dataframes
+
+    Args:ext_npn_measured
+        df(pd.DataFrame): dataframe passed from the ext_measured function
+        dirpath(str): the path to the file where we write data
+        num_workers=mp.cpu_count() (int): num of cpu used
+
+    Returns:
+        df(pd.DataFrame): dataframe contains simulated results
+    """
+
+    results = list()
+    with concurrent.futures.ThreadPoolExecutor(
+        max_workers=num_workers
+    ) as executor:
+        futures_list = list()
+        for j, row in df.iterrows():
+            futures_list.append(
+                executor.submit(
+                    run_sim, dirpath, row["cap"], row["device"], row["temp"]
+                )
+            )
+
+        for future in concurrent.futures.as_completed(futures_list):
+            try:
+                data = future.result()
+                results.append(data)
+            except Exception as exc:
+                print("Test case generated an exception: %s" % (exc))
+
+    sf = glob.glob(f"{dirpath}/cj_simulated/*.csv")
+    sf.sort()
+    # sweeping on all generated cvs files
+    for i in range(0,len(sf)-1,3):
+        sdf = pd.read_csv(
+            sf[i],
+            header=None,
+            delimiter=r"\s+"
+        )
+        sweep = 31 #len(pd.read_csv(glob.glob(f"{dirpath}/cj_measured/*.csv")[1]))
+        new_array = np.empty((sweep, 4))
+        new_array[:, 0] = sdf.iloc[:sweep, 0]
+        new_array[:, 1] = sdf.iloc[:sweep, 1]
+
+        sdf = pd.read_csv(
+            sf[i+1],
+            header=None,
+            delimiter=r"\s+"
+        )
+        new_array[:, 2] = sdf.iloc[:sweep, 1]
+
+        sdf = pd.read_csv(
+            sf[i+2],
+            header=None,
+            delimiter=r"\s+"
+        )
+        new_array[:, 3] = sdf.iloc[:sweep, 1]
+
+    #     for j in range(times):
+    #         new_array[:, (j + 1)] = sdf.iloc[j * sweep : (j + 1) * sweep, 1]
+
+        # Writing final simulated data 1
+        sf[i] = sf[i].replace(".0_cff", "")
+        sdf = pd.DataFrame(new_array)
+        sdf.to_csv(
+            sf[i],
+            index=False
+        )
+        sdf.rename(
+            columns={
+                0: "simulated_volt",
+                1: "simulated_bjt_typical",
+                2: "simulated_bjt_ff",
+                3: "simulated_bjt_ss"
+
+            },
+            inplace=True
+        )
+        sdf.iloc[: , 0:].to_csv(
+            sf[i],
+            index=False
+        )
+    df = pd.DataFrame(results)
+
+    return df
+
+
+
+def error_cal(merged_df: pd.DataFrame, dev_path: str) -> None:
+    """error function calculates the error between measured, simulated data
+
+    Args:
+        merged_df(pd.DataFrame): Dataframe contains devices and csv files
+          which represent measured, simulated data
+        dev_path(str): The path in which we write data
+    """
+
+    # adding error columns to the merged dataframe
+    merged_dfs = list()
+
+    for i in range(len(merged_df)):
+        measured_data = pd.read_csv(merged_df["cj_measured"][i])
+        simulated_data = pd.read_csv(merged_df["cj_measured"][i].replace("measured", "simulated"))
+        # print(simulated_data)
+        # print(measured_data)
+        result_data = measured_data.merge(simulated_data, how="left")
+
+
+        result_data["corner"] = "typical"
+        result_data["device"] = (
+            merged_df["cj_measured"][i]
+            .split("/")[-1]
+            .split("d_")[1]
+            .split("_t")[0]
+        )
+        result_data["temp"] = (
+            merged_df["cj_measured"][i]
+            .split("/")[-1]
+            .split("_")[3]
+            .split("t")[1]
+            .split(".")[0]
+        )
+        exit()
+        result_data["step1_error"] = (
+            np.abs(
+                result_data["simulated_bjt_typical"]
+                - result_data["measured_bjt_typical"]
+            )
+            * 100.0
+            / result_data["measured_bjt_typical"]
+        )
+
+        result_data["step2_error"] = (
+            np.abs(
+                result_data["simulated_bjt_ff"]
+                - result_data["measured_bjt_ff"]
+            )
+            * 100.0
+            / result_data["measured_bjt_ff"]
+        )
+
+        result_data["step3_error"] = (
+            np.abs(
+                result_data["simulated_bjt_ss"]
+                - result_data["measured_bjt_ss"]
+            )
+            * 100.0
+            / result_data["measured_bjt_ss"]
+        )
+
+
+        merged_dfs.append(result_data)
+        merged_out = pd.concat(merged_dfs)
+        merged_out.to_csv(f"{dev_path}/error_analysis.csv", index=False)
+    return None
+
+
 def main():
+    """Main function applies all regression steps"""
 
     # pandas setup
     pd.set_option("display.max_columns", None)
@@ -189,92 +435,75 @@ def main():
     pd.set_option("max_colwidth", None)
     pd.set_option("display.width", 1000)
 
-    corners = ["typical", "ff", "ss"]
-    temps = [25, -40, 175]
-    measure = ["cv", "Vj", "bjt", 31]
-    cv_sim, bjt_vn, bjt_in = measure[0], measure[1], measure[2]
-
-    # vnpn
+    main_regr_dir = "bjt_cj_regr"
+    devices = ["npn", "pnp"]
     npn_devices = [
         "npn_10p00x10p00",
         "npn_05p00x05p00",
         "npn_00p54x16p00",
         "npn_00p54x08p00",
         "npn_00p54x04p00",
-        "npn_00p54x02p00",
+        "npn_00p54x02p00"
     ]
-    npn_start = 0
 
-    for corner in corners:
-        for temp in temps:
-            for device in npn_devices:
-                # Folder structure of measured values
-                dirpath = f"{device}_{cv_sim}_{corner}_T{temp}"
-                if os.path.exists(dirpath) and os.path.isdir(dirpath):
-                    shutil.rmtree(dirpath)
-                os.makedirs(f"{dirpath}/measured_{cv_sim}", exist_ok=False)
-
-                # From xlsx to csv
-                read_file = pd.read_excel(
-                    f"../../180MCU_SPICE_DATA/BJT/{bjt_in}_{cv_sim}_npn.nl_out.xlsx"
-                )
-                read_file.to_csv(f"{dirpath}/{device}.csv", index=False, header=True)
-
-                # Folder structure of simulated values
-                os.makedirs(f"{dirpath}/simulated_{cv_sim}", exist_ok=False)
-                os.makedirs(f"{dirpath}/error_{cv_sim}", exist_ok=False)
-                ext_measured(device, bjt_vn, bjt_in, cv_sim, corner, npn_start, dirpath)
-                ext_simulated(
-                    device, bjt_vn, bjt_in, cv_sim, corner, temp, dirpath, "npn_EBJ"
-                )
-                ext_simulated(
-                    device, bjt_vn, bjt_in, cv_sim, corner, temp, dirpath, "npn_CBJ"
-                )
-                # ext_simulated(device,bjt_vn,bjt_in,cv_sim, corner,temp,dirpath,"npn_CSJ")
-                error_cal(device, bjt_vn, bjt_in, cv_sim, corner, temp, dirpath)
-
-                npn_start = npn_start + 3
-        npn_start = 0
-
-    # vpnp
     pnp_devices = [
         "pnp_10p00x00p42",
         "pnp_05p00x00p42",
         "pnp_10p00x10p00",
-        "pnp_05p00x05p00",
+        "pnp_05p00x05p00"
     ]
-    pnp_start = 0
 
-    for corner in corners:
-        for temp in temps:
-            for device in pnp_devices:
-                # Folder structure of measured values
-                dirpath = f"{device}_{cv_sim}_{corner}_T{temp}"
-                if os.path.exists(dirpath) and os.path.isdir(dirpath):
-                    shutil.rmtree(dirpath)
-                os.makedirs(f"{dirpath}/measured_{cv_sim}", exist_ok=False)
+    for i, dev in enumerate(devices):
+        dev_path = f"{main_regr_dir}/{dev}"
 
-                # From xlsx to csv
-                read_file = pd.read_excel(
-                    f"../../180MCU_SPICE_DATA/BJT/{bjt_in}_{cv_sim}_pnp.nl_out.xlsx"
-                )
-                read_file.to_csv(f"{dirpath}/{device}.csv", index=False, header=True)
+        if os.path.exists(dev_path) :
+            shutil.rmtree(dev_path)
 
-                # Folder structure of simulated values
-                os.makedirs(f"{dirpath}/simulated_{cv_sim}", exist_ok=False)
-                os.makedirs(f"{dirpath}/error_{cv_sim}", exist_ok=False)
-                ext_measured(device, bjt_vn, bjt_in, cv_sim, corner, pnp_start, dirpath)
-                ext_simulated(
-                    device, bjt_vn, bjt_in, cv_sim, corner, temp, dirpath, "pnp_EBJ"
-                )
-                ext_simulated(
-                    device, bjt_vn, bjt_in, cv_sim, corner, temp, dirpath, "pnp_CBJ"
-                )
-                error_cal(device, bjt_vn, bjt_in, cv_sim, corner, temp, dirpath)
+        os.makedirs(f"{dev_path}", exist_ok=False)
 
-                pnp_start = pnp_start + 2
-        pnp_start = 0
+        print("######" * 10)
+        print(f"# Checking Device {dev}")
 
+        cj_data_files = glob.glob(
+            f"../../180MCU_SPICE_DATA/BJT/bjt_cv_{dev}.nl_out.xlsx"
+        )
+        if len(cj_data_files) < 1:
+            print("# Can't find data file for device: {}".format(dev))
+            cj_file = ""
+        else:
+            cj_file = cj_data_files[0]
+        print("# bjt_cj data points file : ", cj_file)
+
+        if cj_file == "":
+            print(f"# No datapoints available for validation for device {dev}")
+            continue
+
+        if dev == "npn":
+            list_dev = npn_devices
+            no_rows = NO_ROWS_NPN
+        elif dev == "pnp":
+            list_dev = pnp_devices
+            no_rows = NO_ROWS_PNP
+
+        if cj_file != "":
+            meas_df = ext_measured(cj_file, dev, list_dev, dev_path, no_rows)
+        else:
+            meas_df = list()
+
+        meas_len = len(
+            pd.read_csv(glob.glob(f"{dev_path}/cj_measured/*.csv")[0])
+        )
+        print(meas_df)
+        sim_df = run_sims(meas_df, dev_path, workers_count)
+
+        # # Merging measured dataframe with the simulated one
+        # merged_df = meas_df.merge(sim_df, on=["device", "temp"], how="left")
+        # # print (merged_df)
+
+         
+        # # passing dataframe to the error_calculation function
+        # # calling error function for creating statistical csv file
+        # error_cal(merged_df, dev_path)
 
 # # ================================================================
 # -------------------------- MAIN --------------------------------
@@ -286,7 +515,7 @@ if __name__ == "__main__":
     arguments = docopt(__doc__, version="comparator: 0.1")
     workers_count = (
         os.cpu_count() * 2
-        if arguments["--num_cores"] == None
+        if arguments["--num_cores"] is None
         else int(arguments["--num_cores"])
     )
